Finish importance sampling

+++ src/utility/rng/xoroshiro128.cpp
@@ -12,7 +12,7 @@ float xoro_128::next()
     s[0] = rotl(s0, 24) ^ s1 ^ (s1 << 16); // a, b
     s[1] = rotl(s1, 37);                   // c
 
-    return float(result * inv_uint64);
+    return result * inv_uint64;
 }
 
 void xoro_128::seed_gen(const uint64_t seed)
